renamed backend.py back to pits_n_giggles.py for sentimental value
