+++ apps/backend/__main__.py
@@ -1,4 +1,4 @@
-from apps.backend.backend import entry_point
+from apps.backend.pits_n_giggles import entry_point
 
 if __name__ == "__main__":
     print("pits_n_giggles starting...\n")
